+++ test/test_helper.rb
@@ -31,7 +31,6 @@ require "minitest/reporters"
 include(StatsD::Instrument::Assertions)
 
 Dir.glob(File.expand_path("../helpers/*.rb", __FILE__)).each { |file| require file }
-ENV["KUBECONFIG"] ||= "#{Dir.home}/.kube/config"
 
 Mocha::Configuration.prevent(:stubbing_method_unnecessarily)
 Mocha::Configuration.prevent(:stubbing_non_existent_method)