+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 *Enhancements*
 - Setting the REVISION environment variable is now optional
+- Defaults KUBECONFIG to `~/.kube/config`
 
 *Bug fixes*
 - Attempting to deploy from a directory that only contains `secrets.ejson` will no longer fail deploy ([#416](https://github.com/Shopify/kubernetes-deploy/pull/416))